Use jenkins-spock 2.1.5 on GitHub instead of 2.1.6-SNAPSHOT
Use snapshot locally, but release on GitHub in order to avoid that Maven
cannot find the snapshot during the build. At the same time,
PublishBuildLogIT will be skipped if "jenkins-spock-2.1.5.jar" is found
on the classpath. I.e., it will be skipped on GitHub, because with that
buggy dependency it will fail. We can remove that restriction after
https://github.com/ExpediaGroup/jenkins-spock/pull/111 was merged or the
corresponding bug fixed otherwise and a new release published.