Skip EmbeddedSpecRunnerTest on JDK 16+
Groovy 2.5.14 does not support Java 16+ class files, but
EmbeddedSpecCompiler.loader is assigned a default GroovyClassLoader
instead of an instance created with a custom CompilerConfiguration.
Therefore, it seems to default to creating class files with a target
byte code version of the currently running JDK instead of what it can
maximally run itself.