Try to stabilise flaky VisJsExampleIT
One key change is to not require the asynchronously loaded target
element (the one to be dragged & dropped) anymore, avoiding error
messages, and also add waiting:

  item4(cache: false, required: false, wait: 5)

Caching stays false, because we need to verify the new coordinates after
DnD.

Another important change is to resize twice, first a dummy resize and
then the one setting the browser to 1024x768 in order to ensure a
more predictable DnD distance. The test page does not seem to refresh
reliably otherwise.