Use stable Chrome version instead of Chromium snapshots in CI builds
Without specifying a version, action 'browser-actions/setup-chrome' by
default installs the latest snapshot (Chrome on Windows and Linux,
Chromium on MacOS). Make sure to use the latest Chrome stable release on
all platforms.

I am not sure if that fixes the flaky DatePickerIT on MacOS, but it
is worth a try.