GitHub CI: re-run failing tests 2x, i.e. run tests max. 3x
In order for this to work, we have to bump Surefire/Failsafe from 2.18.1
to 2.19.1. Any 2.20+ version creates dump files and reports them on the
console, which is really annoying. So we stick with the latest version
which both supports 'rerunFailingTestsCount' on JUnit 4 (Spock 1.3) and
does not create those annoying dump files, but reports native output
correctly on the console.

TODO: Use 3.0.0-M6 when native console problems from M5 have been fixed.